apps/采购订单管理系统/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
pdfplumber==0.11.6
pandas==2.2.3
openpyxl==3.1.5
setuptools
paddlepaddle==3.3.1
paddleocr==2.8.1
opencv-python==4.11.0.86
numpy==1.26.4
